Move FilterRelationships to entity package. Some minor improvements
