+++ pkg/common/entity/helper_test.go
@@ -1,0 +1,39 @@
+package entity
+
+import (
+	"github.com/google/uuid"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestFilterRelationships(t *testing.T) {
+	relationships := []*Relationship{
+		{
+			ID:                  uuid.NullUUID{UUID: uuid.New(), Valid: true},
+			TrustDomainAID:      uuid.New(),
+			TrustDomainBID:      uuid.New(),
+			TrustDomainAConsent: "accepted",
+			TrustDomainBConsent: "denied",
+		},
+		{
+			ID:                  uuid.NullUUID{UUID: uuid.New(), Valid: true},
+			TrustDomainAID:      uuid.New(),
+			TrustDomainBID:      uuid.New(),
+			TrustDomainAConsent: "denied",
+			TrustDomainBConsent: "accepted",
+		},
+	}
+
+	trustDomain := relationships[0].TrustDomainAID
+	status := ConsentStatus("accepted")
+
+	// Call FilterRelationships
+	filtered := FilterRelationships(relationships, &trustDomain, status)
+
+	assert.Equal(t, 1, len(filtered))
+	assert.Equal(t, relationships[0].ID.UUID, filtered[0].ID.UUID)
+	assert.Equal(t, relationships[0].TrustDomainAID, filtered[0].TrustDomainAID)
+	assert.Equal(t, relationships[0].TrustDomainBID, filtered[0].TrustDomainBID)
+	assert.Equal(t, relationships[0].TrustDomainAConsent, filtered[0].TrustDomainAConsent)
+	assert.Equal(t, relationships[0].TrustDomainBConsent, filtered[0].TrustDomainBConsent)
+}
